HW2 выполнил работу над ошибками

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,30 +1,34 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
-import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
 import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
-    private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
-    private AtomicInteger counter= new AtomicInteger(0);
+    private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
+  //  private final Map<Integer, Map<Integer, Meal>> repo = new ConcurrentHashMap<>();
+    private final AtomicInteger counter = new AtomicInteger(0);
 
     {
-       MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> save(meal, meal.getUserId()));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        if (meal == null || meal.getUserId() != userId) return null;
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
+   //         repo.put(userId, repository);
             return meal;
         }
         // handle case: update, but not present in storage
@@ -33,20 +37,24 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        if (repository.get(id).getUserId()!=userId) return false;
+        if (!repository.containsKey(id) || repository.get(id).getUserId() != userId) return false;
         return repository.remove(id) != null;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        if (repository.get(id).getUserId()!=userId) return null;
-        return repository.get(id);
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .findFirst().orElse(null);
     }
 
     @Override
-    public Collection<Meal> getAll(int userId) {
+    public List<Meal> getAll(int userId) {
         Comparator<Meal> comparator = Comparator.comparing(Meal::getDate);
-        return repository.values().stream().filter(meal -> meal.getUserId()==userId).sorted(comparator.reversed()).collect(Collectors.toList());
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .sorted(comparator.reversed())
+                .collect(Collectors.toList());
     }
 }
